refactor: Blog의 이미지를 등록 및 관리 수정
기존 로직을 거둬내고 fileStore를 활용하여 통일

+++ src/main/java/com/skhuedin/skhuedin/controller/api/BlogApiController.java
@@ -2,14 +2,11 @@ package com.skhuedin.skhuedin.controller.api;
 
 import com.skhuedin.skhuedin.controller.response.BasicResponse;
 import com.skhuedin.skhuedin.controller.response.CommonResponse;
-import com.skhuedin.skhuedin.controller.response.ErrorResponse;
 import com.skhuedin.skhuedin.dto.blog.BlogMainResponseDto;
 import com.skhuedin.skhuedin.dto.blog.BlogSaveRequestDto;
-import com.skhuedin.skhuedin.dto.file.FileSaveRequestDto;
 import com.skhuedin.skhuedin.service.BlogService;
-import com.skhuedin.skhuedin.service.FileService;
+import com.skhuedin.skhuedin.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -27,10 +24,8 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.Valid;
-import java.io.File;
 import java.io.IOException;
 import java.security.NoSuchAlgorithmException;
-import java.util.UUID;
 
 @RestController
 @RequestMapping("/api")
@@ -38,49 +33,23 @@ import java.util.UUID;
 public class BlogApiController {
 
     private final BlogService blogService;
-    private final FileService fileService;
-
-    @Value("${resources.storage_location}")
-    private String resourcesLocation;
 
     @PostMapping("blogs")
     @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
     public ResponseEntity<? extends BasicResponse> save(
-            @RequestParam(name = "file", required = false) MultipartFile files,
+            @RequestParam(name = "file", required = false) MultipartFile file,
             @Valid BlogSaveRequestDto requestDto) throws NoSuchAlgorithmException, IOException {
 
-        if (blogService.existsByUserId(requestDto.getUserId())) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body(new ErrorResponse<>("이미 책장이 존재합니다.", "400"));
+        BlogMainResponseDto blogMainResponseDto = blogService.findById(requestDto.getUserId());
+        String email = SecurityUtil.getCurrentEmail().orElse(null);
+        if (email != null && !blogMainResponseDto.getUser().getEmail().equals(email)) {
+            throw new RuntimeException("일치하지 않는 user 정보 입니다.");
         }
 
-        Long saveId;
-        if (files != null && !files.isEmpty()) {
-            String originalFileName = files.getOriginalFilename();
-            String extension = originalFileName.split("\\.")[1];
-            String fileName = UUID.randomUUID() + "." + extension;
-            String savePath = resourcesLocation;
-            if (!new File(savePath).exists()) {
-                new File(savePath).mkdir();
-            }
-            String filePath = savePath + "/" + fileName;
-            files.transferTo(new File(filePath));
-
-            FileSaveRequestDto profile = FileSaveRequestDto
-                    .builder()
-                    .originalName(originalFileName)
-                    .name(fileName)
-                    .path("/profile/" + fileName)
-                    .build();
-
-            Long fileId = fileService.save(profile);
-            saveId = blogService.save(requestDto, fileId);
-        } else {
-            // default image 처리
-            saveId = blogService.save(requestDto, 1L);
-        }
+        Long saveId = blogService.save(requestDto, file);
 
         PageRequest pageRequest = PageRequest.of(0, 10);
+
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(new CommonResponse<>(blogService.findById(saveId, pageRequest)));
     }
@@ -109,36 +78,19 @@ public class BlogApiController {
     @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
     public ResponseEntity<? extends BasicResponse> update(
             @PathVariable("blogId") Long id,
-            @RequestParam(name = "file", required = false) MultipartFile files,
+            @RequestParam(name = "file", required = false) MultipartFile file,
             BlogSaveRequestDto updateDto) throws NoSuchAlgorithmException, IOException {
 
-        Long blogId;
-        if (files != null && !files.isEmpty()) {
-            String originalFileName = files.getOriginalFilename();
-            String extension = originalFileName.split("\\.")[1];
-            String fileName = UUID.randomUUID() + "." + extension;
-            String savePath = resourcesLocation;
-            if (!new File(savePath).exists()) {
-                new File(savePath).mkdir();
-            }
-            String filePath = savePath + "/" + fileName;
-            files.transferTo(new File(filePath));
-
-            FileSaveRequestDto profile = FileSaveRequestDto
-                    .builder()
-                    .originalName(originalFileName)
-                    .name(fileName)
-                    .path("/profile/" + fileName)
-                    .build();
-
-            Long fileId = fileService.save(profile);
-            blogId = blogService.update(id, updateDto, fileId);
-        } else {
-            // default image 처리
-            blogId = blogService.update(id, updateDto, 1L);
+        BlogMainResponseDto blogMainResponseDto = blogService.findById(id);
+        String email = SecurityUtil.getCurrentEmail().orElse(null);
+        if (email != null && !blogMainResponseDto.getUser().getEmail().equals(email)) {
+            throw new RuntimeException("일치하지 않는 user 정보 입니다.");
         }
+
+        blogService.update(updateDto, file);
+
         PageRequest pageRequest = PageRequest.of(0, 10);
-        BlogMainResponseDto responseDto = blogService.findById(blogId, pageRequest);
+        BlogMainResponseDto responseDto = blogService.findById(id, pageRequest);
 
         return ResponseEntity.status(HttpStatus.OK).body(new CommonResponse<>(responseDto));
     }
@@ -146,6 +98,13 @@ public class BlogApiController {
     @DeleteMapping("blogs/{blogId}")
     @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
     public ResponseEntity<? extends BasicResponse> delete(@PathVariable("blogId") Long id) {
+
+        BlogMainResponseDto blogMainResponseDto = blogService.findById(id);
+        String email = SecurityUtil.getCurrentEmail().orElse(null);
+        if (email != null && !blogMainResponseDto.getUser().getEmail().equals(email)) {
+            throw new RuntimeException("일치하지 않는 user 정보 입니다.");
+        }
+
         blogService.delete(id);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
